identify delta table rows and add an option to calculate table diff
