+++ webui/src/constants.ts
@@ -1,3 +1,11 @@
 export const RefTypeBranch = 'branch';
 export const RefTypeCommit = 'commit';
 export const RefTypeTag = 'tag';
+export enum TreeItemType {
+    Object,
+    Prefix,
+    DeltaLakeTable
+}
+export enum TableType {
+    DeltaLake
+}
